fix: ファイル分離
旅行詳細ページの費用概要部分をコンポーネント化しました。
