+++ app/components/AddExpenseForm.tsx
@@ -74,7 +74,7 @@ const AddExpenseForm = ({ tripId }: { tripId: number }) => {
     return (
       <button
         onClick={() => setOpen(true)}
-        className="text-sm text-blue-600 hover:underline cursor-pointer"
+        className="mt-3 text-sm text-blue-600 hover:underline cursor-pointer"
       >
         + 支出を追加
       </button>
